Add k8s test modules #5593
